requirements.txt
flask==2.3.3
numpy
Pillow
gunicorn==21.2.0
python-dotenv
tensorflow-cpu
opencv-python-headless
scipy
